requirements.txt
praw>=7.8.1
python-dotenv>=1.2.1
openai>=1.0.0
requests>=2.32.0
fastapi>=0.109.0
uvicorn>=0.27.0
pytest>=7.4.0
